\name{my.xtics}
\alias{my.xtics}
\title{ Set tic length for axis in grid graphical plot }
\description{
Sets the  tic length for axis in grid graphical plot.
}
\usage{
my.xtics(at, length = 0.5)
}
\arguments{
  \item{at}{gives locations of tics}
  \item{length}{gives length of tics}
}
\details{
Simple wrapper for grid.segments.
}
\value{
Grid graphical object.
}
\author{ David L Duffy }
\keyword{ aplot }
